Added Usage Messages

+++ OS/DiskOS/Programs/factory_reset.lua
@@ -1,3 +1,10 @@
+if select(1,...) == "-?" then
+  printUsage(
+    "factory_reset","Reflashes DiskOS and Reboots"
+  )
+  return
+end
+
 local term = require("C://terminal")
 fs.remove("C://boot.lua")
 for addr=0x15000, 0x17FFF do
